1. Make 2 new files log.h and log.cpp, that include class log, which has 2 methods:   1) warning("message") - show warning message   2) critical("message") - show error and assert simulator
2. All of parrent classes in memory.h (Byte, Byteline, MemoryModel) now derive from log class

3. Correct strange an informationless errors in memory.cpp (like ERROR!)
 
4. Problems:
a) with constant member function:

  hostUInt8 getByteVal( unsigned int num) const;
  Byte getByte( unsigned int num) const;
  ByteLine MemVal::getByteLine( unsigned int index, unsigned int count) const

with our conception of methods log, const function can't change variable - message in methods warning and critical.
So if we used this methdos in const function compilator said error.

Preliminary decision: use previous method of deriving the error in this functions

b) when we use methods of class log in code, assert function write number of the line of this method in file log.cpp but not the number of line where used this method.

Preliminary decision: none

git-svn-id: https://mdsp.googlecode.com/svn/branches@383 77eb06f4-9e05-11de-bf4a-bd5be0e2b6d7